Travels
This is a sample and basic tours and travels project made using java only on netbeans platform. It uses java swing and other components provided by java .
The 1st page is the home page that contains a contact us tab which jumps u to the details page . Then we have the next tab which takes you to the next page of the inquiry form. The details here are saved to database after you click on save. The next page gives you the list of seasons you want to travel in. The next page gives you the details of the countries best suitable for traveling in that season. the last page asks for the inquiry for the tickets and end up with a pop-up box that says you will receive and email soon.